- Añadidos traits para unificar código de las líneas de facturas, albaranes, pedidos, presupuestos y servicios. - Separados los modals de de facturas y albaranes en nuevos archivos en la carpeta view/bloc. - Separadas las líneas de facturas y albaranes de venta en nuevos archivos en la carpeta view/block. - Pequeñas correcciones.
