+++ model/core/linea_pedido_cliente.php
@@ -19,6 +19,8 @@
  */
 namespace FacturaScripts\model;
 
+require_once __DIR__ . '/../../extras/linea_documento_venta.php';
+
 /**
  * Línea de pedido de cliente.
  * 
@@ -27,11 +29,7 @@ namespace FacturaScripts\model;
 class linea_pedido_cliente extends \fs_model
 {
 
-    /**
-     * Clave primaria.
-     * @var integer
-     */
-    public $idlinea;
+    use \linea_documento_venta;
 
     /**
      * ID de la linea relacionada en el presupuesto relacionado,
@@ -53,114 +51,6 @@ class linea_pedido_cliente extends \fs_model
     public $idpresupuesto;
 
     /**
-     * Cantidad del artículo.
-     * @var float
-     */
-    public $cantidad;
-
-    /**
-     * Código del impuesto relacionado.
-     * @var string
-     */
-    public $codimpuesto;
-
-    /**
-     * Descripción del artículo.
-     * @var string
-     */
-    public $descripcion;
-
-    /**
-     * % de descuento.
-     * @var float
-     */
-    public $dtopor;
-
-    /**
-     * % de descuento 2
-     * @var float
-     */
-    public $dtopor2;
-    
-    /**
-     * % de descuento 3
-     * @var float
-     */
-    public $dtopor3;
-    
-    /**
-     * % de descuento 4
-     * @var float
-     */
-    public $dtopor4;
-
-    /**
-     * % de retención IRPF.
-     * @var float
-     */
-    public $irpf;
-
-    /**
-     * % de IVA de la línea, el que corresponde al impuesto.
-     * @var float
-     */
-    public $iva;
-
-    /**
-     * Importe neto sin descuento, es decir, pvpunitario * cantidad.
-     * @var float
-     */
-    public $pvpsindto;
-
-    /**
-     * Importe neto de la línea, sin impuestos.
-     * @var float
-     */
-    public $pvptotal;
-
-    /**
-     * Precio del artículo, una unidad.
-     * @var float
-     */
-    public $pvpunitario;
-
-    /**
-     * % de recargo de equivalencia RE.
-     * @var float
-     */
-    public $recargo;
-
-    /**
-     * Referencia del artículo.
-     * @var string
-     */
-    public $referencia;
-
-    /**
-     * Código de la combinación seleccionada, en el caso de los artículos con atributos.
-     * @var string
-     */
-    public $codcombinacion;
-
-    /**
-     * Posición de la linea en el documento. Cuanto más alto más abajo.
-     * @var integer
-     */
-    public $orden;
-
-    /**
-     * False -> no se muestra la columna cantidad al imprimir.
-     * @var boolean
-     */
-    public $mostrar_cantidad;
-
-    /**
-     * False -> no se muestran las columnas precio, descuento, impuestos y total al imprimir.
-     * @var boolean
-     */
-    public $mostrar_precio;
-
-    /**
      * Listado de pedidos.
      * @var array
      */
@@ -175,72 +65,16 @@ class linea_pedido_cliente extends \fs_model
         }
 
         if ($data) {
-            $this->idlinea = $this->intval($data['idlinea']);
+            $this->load_data_trait($data);
             $this->idlineapresupuesto = $this->intval($data['idlineapresupuesto']);
             $this->idpedido = $this->intval($data['idpedido']);
             $this->idpresupuesto = $this->intval($data['idpresupuesto']);
-            $this->cantidad = floatval($data['cantidad']);
-            $this->codcombinacion = $data['codcombinacion'];
-            $this->codimpuesto = $data['codimpuesto'];
-            $this->descripcion = $data['descripcion'];
-            $this->dtopor = floatval($data['dtopor']);
-            $this->dtopor2 = floatval($data['dtopor2']);
-            $this->dtopor3 = floatval($data['dtopor3']);
-            $this->dtopor4 = floatval($data['dtopor4']);
-            $this->irpf = floatval($data['irpf']);
-            $this->iva = floatval($data['iva']);
-            $this->pvpsindto = floatval($data['pvpsindto']);
-            $this->pvptotal = floatval($data['pvptotal']);
-            $this->pvpunitario = floatval($data['pvpunitario']);
-            $this->recargo = floatval($data['recargo']);
-            $this->referencia = $data['referencia'];
-            $this->orden = intval($data['orden']);
-            $this->mostrar_cantidad = $this->str2bool($data['mostrar_cantidad']);
-            $this->mostrar_precio = $this->str2bool($data['mostrar_precio']);
         } else {
-            $this->idlinea = NULL;
+            $this->clear_trait();
             $this->idlineapresupuesto = NULL;
             $this->idpedido = NULL;
             $this->idpresupuesto = NULL;
-            $this->cantidad = 0.0;
-            $this->codimpuesto = NULL;
-            $this->codcombinacion = NULL;
-            $this->descripcion = '';
-            $this->dtopor = 0.0;
-            $this->dtopor2 = 0.0;
-            $this->dtopor3 = 0.0;
-            $this->dtopor4 = 0.0;
-            $this->irpf = 0.0;
-            $this->iva = 0.0;
-            $this->mostrar_cantidad = TRUE;
-            $this->mostrar_precio = TRUE;
-            $this->orden = 0;
-            $this->pvpsindto = 0.0;
-            $this->pvptotal = 0.0;
-            $this->pvpunitario = 0.0;
-            $this->recargo = 0.0;
-            $this->referencia = NULL;
         }
-    }
-
-    protected function install()
-    {
-        return '';
-    }
-
-    public function pvp_iva()
-    {
-        return $this->pvpunitario * (100 + $this->iva) / 100;
-    }
-
-    public function total_iva()
-    {
-        return $this->pvptotal * (100 + $this->iva - $this->irpf + $this->recargo) / 100;
-    }
-
-    public function descripcion()
-    {
-        return nl2br($this->descripcion);
     }
 
     public function show_codigo()
@@ -314,15 +148,6 @@ class linea_pedido_cliente extends \fs_model
         return 'index.php?page=ventas_pedido&id=' . $this->idpedido;
     }
 
-    public function articulo_url()
-    {
-        if (is_null($this->referencia) OR $this->referencia == '') {
-            return "index.php?page=ventas_articulos";
-        }
-
-        return "index.php?page=ventas_articulo&ref=" . urlencode($this->referencia);
-    }
-
     public function exists()
     {
         if (is_null($this->idlinea)) {
@@ -330,24 +155,6 @@ class linea_pedido_cliente extends \fs_model
         }
 
         return $this->db->select("SELECT * FROM " . $this->table_name . " WHERE idlinea = " . $this->var2str($this->idlinea) . ";");
-    }
-
-    public function test()
-    {
-        $this->descripcion = $this->no_html($this->descripcion);
-        $totalsindto = $this->pvpunitario * $this->cantidad;
-        $dto_due = (1-((1-$this->dtopor/100)*(1-$this->dtopor2/100)*(1-$this->dtopor3/100)*(1-$this->dtopor4/100)))*100;
-        $total = $totalsindto * (1 - $dto_due / 100);
-
-        if (!$this->floatcmp($this->pvptotal, $total, FS_NF0, TRUE)) {
-            $this->new_error_msg("Error en el valor de pvptotal de la línea " . $this->referencia . " del " . FS_PEDIDO . ". Valor correcto: " . $total . " y se recibe " . $this->pvptotal);
-            return FALSE;
-        } else if (!$this->floatcmp($this->pvpsindto, $totalsindto, FS_NF0, TRUE)) {
-            $this->new_error_msg("Error en el valor de pvpsindto de la línea " . $this->referencia . " del " . FS_PEDIDO . ". Valor correcto: " . $totalsindto . " y se recibe " . $this->pvptotal);
-            return FALSE;
-        }
-
-        return TRUE;
     }
 
     public function save()
