+++ view/block/compras_factura_modals.html
@@ -1,0 +1,129 @@
+<div class="modal fade" id="modal_imprimir">
+   <div class="modal-dialog">
+      <div class="modal-content">
+         <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+            <h4 class="modal-title">Imprimir factura</h4>
+         </div>
+         <div class="modal-body">
+            {loop="$fsc->extensions"}
+               {if="$value->type=='pdf'"}
+               <a href="index.php?page={$value->from}{$value->params}&id={$fsc->factura->idfactura}" target="_blank" class="btn btn-block btn-default">
+                  <span class="glyphicon glyphicon-print"></span> &nbsp; {$value->text}
+               </a>
+               {/if}
+            {/loop}
+         </div>
+         <div class="modal-footer">
+            <a href="index.php?page=admin_empresa#impresion" target="_blank">
+               <span class="glyphicon glyphicon-wrench"></span>&nbsp; Opciones de impresión
+            </a>
+         </div>
+      </div>
+   </div>
+</div>
+
+<form action="{$fsc->url()}&pagada=TRUE" method="post" class="form">
+   <div class="modal fade" id="modal_pagar">
+      <div class="modal-dialog modal-sm">
+         <div class="modal-content">
+            <div class="modal-header">
+               <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+               <h4 class="modal-title">Marcar factura como pagada</h4>
+            </div>
+            <div class="modal-body">
+               <div class="form-group">
+                  Fecha del pago:
+                  <div class="input-group">
+                     <span class="input-group-addon">
+                        <span class="glyphicon glyphicon-calendar"></span>
+                     </span>
+                     <input type="text" name="fpagada" value="{$fsc->today()}" class="form-control datepicker" required="" autocomplete="off"/>
+                  </div>
+                  {if="$fsc->empresa->contintegrada"}
+                  <p class="help-block">Se generará un asiento de pago.</p>
+                  {/if}
+               </div>
+               <div class="text-right">
+                  <button type="submit" class="btn btn-sm btn-primary">
+                     <span class="glyphicon glyphicon-floppy-disk"></span>&nbsp; Guardar
+                  </button>
+               </div>
+            </div>
+         </div>
+      </div>
+   </div>
+</form>
+
+<form action="{$fsc->url()}" method="post" class="form">
+   <input type="hidden" name="anular" value="TRUE"/>
+   <div class="modal fade" id="modal_eliminar">
+      <div class="modal-dialog">
+         <div class="modal-content">
+            <div class="modal-header">
+               <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+               <h4 class="modal-title">¿Quiere eliminar o anular esta factura?</h4>
+               {if="!in_array('editar_facturas', $GLOBALS['plugins'])"}
+               <p class="help-block">
+                  Puedes editar facturas usando el plugin
+                  <a href="https://www.facturascripts.com/plugin/editar_facturas" target="_blank">editar facturas</a>.
+               </p>
+               {/if}
+            </div>
+            <div class="modal-body bg-warning">
+               {if="$fsc->factura->idasiento"}
+                  Si decide <b>eliminar</b>, hay asociado un asiento contable que será eliminado
+                  junto con la factura. Además, si no hay asociado un {#FS_ALBARAN#} o {#FS_ALBARANES#},
+                  se restaurará el stock de los artículos.
+               {else}
+                  Si decide <b>eliminar</b>, se restaurará el stock de los artículos si no hay asociado un
+                  {#FS_ALBARAN#} o {#FS_ALBARANES#}.
+               {/if}
+               {if="!$fsc->factura->anulada"}
+                  <br/><br/>
+                  {if="$fsc->empresa->codpais!='ESP'"}
+                  Si decide <b>anular</b> la factura se restaurará el stock, a menos que haya asociado
+                  un {#FS_ALBARAN#} o {#FS_ALBARANES#}.
+                  <br/><br/>
+                  {/if}
+                  Y si decide generarar una <b>{#FS_FACTURA_RECTIFICATIVA#}</b>, se anulará esta
+                  y se restaurará el stock de los artículos, aunque primero debe elegir la serie
+                  para la nueva {#FS_FACTURA_RECTIFICATIVA#}:
+                  <div class="form-group">
+                     <select name="codserie" class="form-control">
+                     {loop="$fsc->serie->all()"}
+                        {if="$value->codserie=='R'"}
+                        <option value="{$value->codserie}" selected="">{$value->descripcion}</option>
+                        {else}
+                        <option value="{$value->codserie}">{$value->descripcion}</option>
+                        {/if}
+                     {/loop}
+                     </select>
+                  </div>
+                  <div class="form-group">
+                     <textarea name="motivo" class="form-control" placeholder="Motivo de la anulación"></textarea>
+                  </div>
+               {/if}
+            </div>
+            <div class="modal-footer">
+               <a class="btn btn-sm btn-danger pull-left" href="{$fsc->ppage->url()}&delete={$fsc->factura->idfactura}">
+                  <span class="glyphicon glyphicon-trash"></span>
+                  <span class='hidden-xs'>&nbsp;Eliminar</span>
+               </a>
+               {if="!$fsc->factura->anulada"}
+               <div class="btn-group">
+                  {if="$fsc->empresa->codpais!='ESP'"}
+                  <button type="submit" name="rectificativa" value="FALSE" class="btn btn-sm btn-warning">
+                     Anular
+                  </button>
+                  {/if}
+                  <button type="submit" name="rectificativa" value="TRUE" class="btn btn-sm btn-warning text-capitalize">
+                     {#FS_FACTURA_RECTIFICATIVA#}
+                  </button>
+               </div>
+               {/if}
+            </div>
+         </div>
+      </div>
+   </div>
+</form>
