+++ view/block/nueva_compra.html
@@ -130,7 +130,7 @@
         </ul>
         <div class="tab-content">
             <div role="tabpanel" class="tab-pane active" id="lineas">
-                {include="block/table_lineas_nueva_compra"}
+                {include="block/nueva_compra_lineas"}
             </div>
             <div role="tabpanel" class="tab-pane" id="detalles">
                 <div class="container-fluid" style="margin-top: 10px;">
