+++ view/block/nueva_compra_lineas.html
