+++ view/block/nueva_venta.html
@@ -148,7 +148,7 @@
         </ul>
         <div class="tab-content">
             <div role="tabpanel" class="tab-pane active" id="lineas">
-                {include="block/table_lineas_nueva_venta"}
+                {include="block/nueva_venta_lineas"}
             </div>
             <div role="tabpanel" class="tab-pane" id="detalles">
                 <div class="container-fluid" style="margin-top: 10px;">
